Script is ready for merging.
